requirements.txt
Django==1.3
django-autoslug==1.4.1
django-extensions==0.6
django-tagging==0.3.1
sorl-thumbnail==11.05.2
psycopg2==2.4.2
wsgiref==0.1.2
